[tsp-client] Fixes for yargs update + more testing (#8758)
* refactor commands

* add test files

* add sync tests

* rename test file

* clean up test

* fix update command tsp-location.yaml logic

* add update tests

* init fix and test

---------
