+++ tools/tsp-client/test/examples/sdk/contosowidgetmanager/contosowidgetmanager-rest/tsp-location.yaml
@@ -1,5 +1,5 @@
 directory: specification/contosowidgetmanager/Contoso.WidgetManager
 commit: db63bea839f5648462c94e685d5cc96f8e8b38ba
 repo: Azure/azure-rest-api-specs
-additionalDirectories:
-- specification/contosowidgetmanager/Contoso.WidgetManager.Shared/
+additionalDirectories: 
+- specification/contosowidgetmanager/Contoso.WidgetManager.Shared
