+++ tools/tsp-client/test/examples/sdk/local-spec-sdk/tsp-location.yaml
@@ -1,0 +1,5 @@
+directory: tools/tsp-client/test/examples/specification/contosowidgetmanager/Contoso.WidgetManager
+commit: db63bea839f5648462c94e685d5cc96f8e8b38ba
+repo: Azure/azure-sdk-tools
+additionalDirectories:
+  - tools/tsp-client/test/examples/specification/contosowidgetmanager/Contoso.WidgetManager.Shared/
