+++ tools/tsp-client/test/examples/specification/contosowidgetmanager/Contoso.WidgetManager.Shared/main.tsp
@@ -1,0 +1,8 @@
+@doc("Faked shared model")
+model FakedSharedModel {
+  @doc("The tag.")
+  tag: string;
+
+  @doc("The created date.")
+  createdAt: utcDateTime;
+}
