+++ tools/tsp-client/test/examples/specification/contosowidgetmanager/Contoso.WidgetManager/examples/2022-11-01-preview/Widgets_CreateOrUpdateWidgetSample.json
@@ -1,0 +1,37 @@
+{
+  "title": "Widgets_CreateOrUpdateWidget",
+  "operationId": "Widgets_CreateOrUpdateWidget",
+  "parameters": {
+    "widgetName": "name1",
+    "api-version": "2022-11-01-preview",
+    "resource": {
+      "manufacturerId": "manufacturer id1",
+      "sharedModel": {
+        "tag": "tag1",
+        "createdAt": "2023-01-09T02:12:25.689Z"
+      }
+    }
+  },
+  "responses": {
+    "200": {
+      "body": {
+        "name": "name1",
+        "manufacturerId": "manufacturer id1",
+        "sharedModel": {
+          "tag": "tag1",
+          "createdAt": "2023-01-09T02:12:25.689Z"
+        }
+      }
+    },
+    "201": {
+      "body": {
+        "name": "name1",
+        "manufacturerId": "manufacturer id1",
+        "sharedModel": {
+          "tag": "tag1",
+          "createdAt": "2023-01-09T02:12:25.689Z"
+        }
+      }
+    }
+  }
+}
