+++ tools/tsp-client/test/examples/specification/contosowidgetmanager/Contoso.WidgetManager/examples/2022-11-01-preview/Widgets_DeleteWidgetSample.json
@@ -1,0 +1,29 @@
+{
+  "operationId": "Widgets_DeleteWidget",
+  "title": "Delete widget by widget name using long-running operation.",
+  "parameters": {
+    "api-version": "2022-11-01-preview",
+    "widgetName": "searchbox"
+  },
+  "responses": {
+    "202": {
+      "headers": {
+        "location": "https://contosowidgetmanager.azure.com/operations/00000000-0000-0000-0000-000000000123/result?api-version=2022-11-01-preview",
+        "operation-location": "https://contosowidgetmanager.azure.com/operations/00000000-0000-0000-0000-000000000123?api-version=2022-11-01-preview"
+      },
+      "body": {
+        "id": "id1",
+        "status": "deleted"
+      }
+    },
+    "default": {
+      "body": {
+        "error": {
+          "code": "Error code",
+          "message": "Error message",
+          "details": []
+        }
+      }
+    }
+  }
+}
