+++ tools/tsp-client/test/examples/specification/contosowidgetmanager/Contoso.WidgetManager/examples/2022-11-01-preview/Widgets_GetWidgetOperationStatusSample.json
@@ -1,0 +1,45 @@
+{
+  "title": "Widgets_GetWidgetOperationStatus",
+  "operationId": "Widgets_GetWidgetOperationStatus",
+  "parameters": {
+    "widgetName": "name1",
+    "operationId": "opreation id1",
+    "api-version": "2022-11-01-preview"
+  },
+  "responses": {
+    "200": {
+      "body": {
+        "id": "opreation id1",
+        "status": "InProgress",
+        "error": {
+          "code": "Error code",
+          "message": "Error message",
+          "target": "op1",
+          "details": [
+            {
+              "code": "code1",
+              "message": "message1",
+              "target": "op1",
+              "details": [],
+              "innererror": {
+                "code": "code1"
+              }
+            }
+          ],
+          "innererror": {
+            "code": "code1"
+          }
+        },
+        "result": {
+          "name": "bingsearch",
+          "manufacturerId": "manufacturer Id1",
+          "sharedModel": {
+            "tag": "tag1",
+            "createdAt": "2023-01-09T02:12:25.689Z"
+          }
+        },
+        "widgetName": "rfazvwnfwwomiwrh"
+      }
+    }
+  }
+}
