+++ tools/tsp-client/test/examples/specification/contosowidgetmanager/Contoso.WidgetManager/examples/2022-11-01-preview/Widgets_GetWidgetSample.json
@@ -1,0 +1,25 @@
+{
+  "operationId": "Widgets_GetWidget",
+  "title": "Get widget by widget name.",
+  "parameters": {
+    "api-version": "2022-11-01-preview",
+    "widgetName": "searchbox"
+  },
+  "responses": {
+    "200": {
+      "body": {
+        "name": "bingsearch",
+        "manufacturerId": "a-22-01"
+      }
+    },
+    "default": {
+      "body": {
+        "error": {
+          "code": "Error code",
+          "message": "Error message",
+          "details": []
+        }
+      }
+    }
+  }
+}
