+++ tools/tsp-client/test/examples/specification/contosowidgetmanager/Contoso.WidgetManager/examples/2022-11-01-preview/Widgets_ListWidgetsSample.json
@@ -1,0 +1,27 @@
+{
+  "title": "Widgets_ListWidgets",
+  "operationId": "Widgets_ListWidgets",
+  "parameters": {
+    "top": 8,
+    "skip": 15,
+    "maxpagesize": 27,
+    "api-version": "2022-11-01-preview"
+  },
+  "responses": {
+    "200": {
+      "body": {
+        "value": [
+          {
+            "name": "bingsearch",
+            "manufacturerId": "manufacturer Id1",
+            "sharedModel": {
+              "tag": "tag1",
+              "createdAt": "2023-01-09T02:12:25.689Z"
+            }
+          }
+        ],
+        "nextLink": "https://microsoft.com/a"
+      }
+    }
+  }
+}
