+++ tools/tsp-client/test/examples/specification/contosowidgetmanager/Contoso.WidgetManager/main.tsp
@@ -1,0 +1,63 @@
+import "@typespec/http";
+import "@typespec/rest";
+import "@typespec/versioning";
+import "@azure-tools/typespec-azure-core";
+import "../Contoso.WidgetManager.Shared";
+
+using TypeSpec.Http;
+using TypeSpec.Rest;
+using TypeSpec.Versioning;
+using Azure.Core;
+
+@useAuth(AadOauth2Auth<["https://contoso.azure.com/.default"]>)
+@service({
+  title: "Contoso Widget Manager",
+})
+@versioned(Contoso.WidgetManager.Versions)
+namespace Azure.Contoso.WidgetManager;
+
+@doc("Versions info.")
+enum Versions {
+  @doc("The 2022-11-01-preview version.")
+  @useDependency(Azure.Core.Versions.v1_0_Preview_1)
+  v2022_11_01_Preview: "2022-11-01-preview",
+}
+
+@doc("A widget.")
+@resource("widgets")
+model WidgetSuite {
+  @key("widgetName")
+  @doc("The widget name.")
+  @visibility("read")
+  name: string;
+
+  @doc("The ID of the widget's manufacturer.")
+  manufacturerId: string;
+
+  @doc("The faked shared model.")
+  sharedModel?: FakedSharedModel;
+}
+
+interface Widgets {
+  @doc("Fetch a Widget by name.")
+  getWidget is ResourceRead<WidgetSuite>;
+
+  @doc("Gets status of a Widget operation.")
+  getWidgetOperationStatus is GetResourceOperationStatus<WidgetSuite>;
+
+  @doc("Creates or updates a Widget asynchronously.")
+  @pollingOperation(Widgets.getWidgetOperationStatus)
+  createOrUpdateWidget is StandardResourceOperations.LongRunningResourceCreateOrUpdate<WidgetSuite>;
+
+  @doc("Delete a Widget asynchronously.")
+  @pollingOperation(Widgets.getWidgetOperationStatus)
+  deleteWidget is LongRunningResourceDelete<WidgetSuite>;
+
+  @doc("List Widget resources")
+  listWidgets is ResourceList<
+    WidgetSuite,
+    {
+      parameters: StandardListQueryParameters;
+    }
+  >;
+}
